fix: Move CompactFareContracts out of stacks-hiearchy
And move the details-functions that it uses
