+++ src/modules/fare-contracts/sections/FareContractInfoDetailsSectionItem.tsx
@@ -1,5 +1,4 @@
 import {
-  findReferenceDataById,
   PreassignedFareProduct,
   SupplementProduct,
   FareZone,
@@ -7,15 +6,12 @@ import {
   UserProfile,
 } from '@atb/modules/configuration';
 import {StyleSheet} from '@atb/theme';
-import {getLastUsedAccess} from '@atb/modules/ticketing';
 import {FareContractType} from '@atb-as/utils';
 import {FareContractTexts, useTranslation} from '@atb/translations';
 import React from 'react';
 import {View} from 'react-native';
 import {
-  getValidityStatus,
   isValidFareContract,
-  mapToUserProfilesWithCount,
   useFareZoneSummary,
   ValidityStatus,
 } from '../utils';
@@ -23,14 +19,10 @@ import {FareContractDetailItem} from '../components/FareContractDetailItem';
 import {InspectionSymbol} from '../components/InspectionSymbol';
 import {getTransportModeText} from '@atb/components/transportation-modes';
 import {SectionItemProps, useSectionItem} from '@atb/components/sections';
-import {getAccesses} from '@atb-as/utils';
-import {isDefined} from '@atb/utils/presence';
 import {
-  arrayMapUniqueWithCount,
   toCountAndName,
   UniqueWithCount,
 } from '@atb/utils/array-map-unique-with-count';
-import {getBaggageProducts} from '../get-baggage-products';
 
 export type FareContractInfoProps = {
   status: ValidityStatus;
@@ -130,75 +122,6 @@ export const FareContractInfoDetailsSectionItem = ({
       </View>
     </View>
   );
-};
-
-export const getFareContractInfoDetails = (
-  fareContract: FareContractType,
-  now: number,
-  fareZones: FareZone[],
-  userProfiles: UserProfile[],
-  preassignedFareProducts: PreassignedFareProduct[],
-  supplementProducts: SupplementProduct[],
-): FareContractInfoDetailsProps => {
-  const {
-    endDateTime,
-    fareProductRef: productRef,
-    fareZoneRefs,
-  } = fareContract.travelRights[0];
-  let validTo = endDateTime.getTime();
-  const validityStatus = getValidityStatus(now, fareContract);
-
-  const firstZone = fareZoneRefs?.[0];
-  const lastZone = fareZoneRefs?.slice(-1)?.[0];
-  const fromFareZone = firstZone
-    ? findReferenceDataById(fareZones, firstZone)
-    : undefined;
-  const toFareZone = lastZone
-    ? findReferenceDataById(fareZones, lastZone)
-    : undefined;
-  const preassignedFareProduct = findReferenceDataById(
-    preassignedFareProducts,
-    productRef,
-  );
-  const userProfilesWithCount = mapToUserProfilesWithCount(
-    fareContract.travelRights.map((tr) => tr.userProfileRef).filter(isDefined),
-    userProfiles,
-  );
-
-  const productsInFareContract = fareContract.travelRights
-    .map((tr) =>
-      findReferenceDataById(preassignedFareProducts, tr.fareProductRef),
-    )
-    .filter(isDefined);
-
-  const baggageProducts = getBaggageProducts(
-    productsInFareContract,
-    supplementProducts,
-  );
-
-  const baggageProductsWithCount = arrayMapUniqueWithCount(
-    baggageProducts,
-    (a, b) => a.id === b.id,
-  );
-
-  const flattenedAccesses = getAccesses(fareContract);
-  if (flattenedAccesses) {
-    const {usedAccesses} = flattenedAccesses;
-    const {validTo: usedAccessValidTo} = getLastUsedAccess(now, usedAccesses);
-    if (usedAccessValidTo) validTo = usedAccessValidTo;
-  }
-
-  return {
-    preassignedFareProduct: preassignedFareProduct,
-    fromFareZone: fromFareZone,
-    toFareZone: toFareZone,
-    userProfilesWithCount,
-    baggageProductsWithCount,
-    status: validityStatus,
-    now: now,
-    validTo: validTo,
-    fareContract,
-  };
 };
 
 const useStyles = StyleSheet.createThemeHook((theme) => ({
